style: rename filter utility files to kebab-case per coding guidelines
